Initial proposal for how to implement
Main question is how to do a recurring timer to swap the zem every 1 week if the zone has stayed up that long.  I think checking every 1 hour like we do for weather wouldn't be too bad.  If there's too much overhead adding a new timer that runs in terms in milliseconds, we could piggyback off an existing one like the weather which runs every 1 minute.