Add /consent #raid support
Enable support for granting consent automatically to all raid members
using a new column in the `raid_members` table.

/consent #raid on: Enables auto-raid consent
/consent #raid off: Disables auto-raid consent
/consent #raid: Toggles auto-raid consent

Requires the addition of an `is_consent` column to the `raid_members`
player database table:
`alter table raid_members add is_consent tinyint(1) not null default 0;`

The commands control flags in the client which are used when
joining a raid to set the is_consent value and also trigger updates
to the raid database with a group update, supporting the granting and
revoking of consent across zones.